refactor: seperate irreducible_sector out of symmetry_rotation
fix a symmetry bug

fix default value

rebase develop and small fix

refactor rotation funcs

pointer-type interface for rotation funcs
