+++ source/module_ri/exx_symmetry/CMakeLists.txt
@@ -1,0 +1,21 @@
+if (ENABLE_LIBRI)
+
+    list(APPEND objects
+        irreducible_sector.cpp
+        irreducible_sector_bvk.cpp
+        symmetry_rotation.cpp
+    )
+    add_library(
+        exx_symmetry
+        OBJECT
+        ${objects}
+    )
+    if(BUILD_TESTING)
+        add_subdirectory(test)
+    endif()
+    
+    if(ENABLE_COVERAGE)
+        add_coverage(exx_symmetry)
+    endif()
+
+  endif()
