+++ source/module_ri/exx_symmetry/symmetry_rotation.h
@@ -1,18 +1,10 @@
 #pragma once
-#include "module_base/abfs-vector3_order.h"
-#include "module_base/complexmatrix.h"
-#include "module_base/matrix3.h"
-#include <vector>
-#include <map>
-#include <set>
-#include "module_basis/module_ao/parallel_2d.h"
-#include "module_cell/unitcell.h" 
-#include "module_cell/module_symmetry/symmetry.h"
-#include "module_cell/klist.h"
+#include "irreducible_sector.h"
+#include "module_basis/module_ao/parallel_orbitals.h"
 #include <RI/global/Tensor.h>
 #include "module_hamilt_lcao/module_hcontainer/hcontainer.h"
 #include "module_cell/module_neighbor/sltk_grid_driver.h"
- 
+
 namespace ModuleSymmetry
 {
     using Tap = std::pair<int, int>;
@@ -20,49 +12,32 @@ namespace ModuleSymmetry
     using TapR = std::pair<Tap, TC>;
     using TCdouble = Abfs::Vector3_Order<double>;
 
-    struct ap_less_func
-    {
-        bool operator()(const Tap& lhs, const Tap& rhs) const
-        {
-            if (lhs.first < rhs.first)return true;
-            else if (lhs.first > rhs.first)return false;
-            else return lhs.second < rhs.second;
-        }
-    };
-    struct apR_less_func
-    {
-        bool operator()(const TapR& lhs, const TapR& rhs) const
-        {
-            if (lhs.first < rhs.first)return true;
-            else if (lhs.first > rhs.first)return false;
-            else return lhs.second < rhs.second;
-        }
-    };
-    struct len_less_func
-    {
-        int norm2(const TC& R)const
-        {
-            return R[0] * R[0] + R[1] * R[1] + R[2] * R[2];
-        }
-        bool operator()(const TC& lhs, const TC& rhs) const
-        {
-            if (norm2(lhs) < norm2(rhs))return true;
-            else if (norm2(lhs) > norm2(rhs))return false;
-            else return lhs < rhs;
-        }
-    };
-
     class Symmetry_rotation
     {
     public:
         Symmetry_rotation() {};
         ~Symmetry_rotation() {};
-        //--------------------------------------------------------------------------------
-        /// functions  to contruct rotation matrix in AO-representation
 
-        /// The top-level calculation interface of this class. calculate the rotation matrix in AO representation: M
-        /// only need once call in each ion step (decided by the configuration)
-        /// @param kstars  equal k points to each ibz-kpont, corresponding to a certain symmetry operations. 
+        //--------------------------------------------------------------------------------
+        // getters
+        const std::map<Tap, std::set<TC>>& get_irreducible_sector()const { return this->irs_.get_irreducible_sector(); }
+        void find_irreducible_sector(const Symmetry& symm, const Atom* atoms, const Statistics& st,
+            const std::vector<TC>& Rs, const TC& period, const Lattice& lat)
+        {
+            this->irs_.find_irreducible_sector(symm, atoms, st, Rs, period, lat);
+        }
+        TCdouble get_return_lattice(const Symmetry& symm,
+            const ModuleBase::Matrix3& gmatd, const TCdouble gtransd,
+            const TCdouble& posd_a1, const TCdouble& posd_a2)const
+        {
+            return this->irs_.get_return_lattice(symm, gmatd, gtransd, posd_a1, posd_a2);
+        }
+        //--------------------------------------------------------------------------------
+/// functions  to contruct rotation matrix in AO-representation
+
+/// The top-level calculation interface of this class. calculate the rotation matrix in AO representation: M
+/// only need once call in each ion step (decided by the configuration)
+/// @param kstars  equal k points to each ibz-kpont, corresponding to a certain symmetry operations. 
         void cal_Ms(const K_Vectors& kv,
             //const std::vector<std::map<int, TCdouble>>& kstars,
             const UnitCell& ucell, const Parallel_2D& pv);
@@ -91,14 +66,6 @@ namespace ModuleSymmetry
         /// T_mm' = [c^\dagger D c]_mm', the rotation matrix in the representation of real sphere harmonics
         void cal_rotmat_Slm(const ModuleBase::Matrix3* gmatc, const int lmax);
 
-        /// Perfoming {R|t} to atom position r in the R=0 lattice, we get Rr+t, which may get out of R=0 lattice, 
-        /// whose image in R=0 lattice is r'=Rr+t-O. This function is to get O for each atom and each symmetry operation.
-        /// the range of direct position is [-0.5, 0.5).
-        TCdouble get_return_lattice(const Symmetry& symm,
-            const ModuleBase::Matrix3& gmatd, const TCdouble gtransd,
-            const TCdouble& posd_a1, const TCdouble& posd_a2)const;
-        void get_return_lattice_all(const Symmetry& symm, const Atom* atoms, const Statistics& st);
-
         /// set a block matrix onto a 2d-parallelized matrix(col-maj), at the position (starti, startj) 
         /// if trans=true, the block matrix is transposed before setting
         void set_block_to_mat2d(const int starti, const int startj, const ModuleBase::ComplexMatrix& block,
@@ -111,19 +78,7 @@ namespace ModuleSymmetry
         std::vector<std::complex<double>> contruct_2d_rot_mat_ao(const Symmetry& symm, const Atom* atoms, const Statistics& cell_st,
             const TCdouble& kvec_d_ibz, int isym, const Parallel_2D& pv) const;
 
-        ModuleBase::Matrix3 direct_to_cartesian(const ModuleBase::Matrix3& d, const ModuleBase::Matrix3& latvec)const;
-
         std::vector<std::vector<ModuleBase::ComplexMatrix>>& get_rotmat_Slm() { return this->rotmat_Slm_; }
-
-        //--------------------------------------------------------------------------------
-        /// The main function to find irreducible sector: {abR}
-        void find_irreducible_sector(const Symmetry& symm, const Atom* atoms, const Statistics& st,
-            const std::vector<TC>& Rs, const TC& period, const Lattice& lat);
-        std::vector<TC> get_Rs_from_BvK(const K_Vectors& kv)const;
-        std::vector<TC> get_Rs_from_adjacent_list(const UnitCell& ucell, Grid_Driver& gd, const Parallel_Orbitals& pv)const;
-        const std::map<Tap, std::set<TC>>& get_irreducible_sector()const { return this->irreducible_sector_; }
-        // const std::map<int, std::set<std::pair<int, TC>>> convirt_irreducible_sector() {};
-        //--------------------------------------------------------------------------------
 
         //--------------------------------------------------------------------------------
         /// The main functions to rotate matrices
@@ -151,47 +106,26 @@ namespace ModuleSymmetry
         //--------------------------------------------------------------------------------
 
     private:
-        int round2int(const double x)const;
-
         //--------------------------------------------------------------------------------
-        /// The sub functions to find irreducible sector: {abR}
-
-        /// gauge='L' means H(R)=<R|H|0>; gauge='R' means H(R)=<0|H|R>
-        /// gauge='L': R'=R+O_1-O_2; gauge='R': R'=R+O_2-O_1
-        TC rotate_R_by_formula(const Symmetry& symm, const int isym, const int iat1, const int iat2, const TC& R, const char gauge = 'R')const;
-        TapR rotate_R_by_formula(const Symmetry& symm, const int isym, const TapR& iapR, const char gauge = 'R')const;
-        /// gauge='L': tau_a + R - tau_b; gauge='R': tau_a - tau_b - R (direct)
-        TCdouble get_aRb_direct(const Atom* atoms, const Statistics& st, const int iat1, const int iat2, const TC& R, const char gauge = 'R')const;
-        TCdouble get_aRb_direct(const Atom* atoms, const Statistics& st, const int iat1, const int iat2, const TCdouble& R, const char gauge = 'R')const;
-
-        // /// find the irreducible atom pairs
-        // /// algorithm 1: the way finding irreducible k-points
-        // void find_irreducible_atom_pairs(const Symmetry& symm);
-        // /// algorithm 2: taking out atom pairs from the initial set
-        // void find_irreducible_atom_pairs_set(const Symmetry& symm);
-        // /// double check between the two algorithms
-        // void test_irreducible_atom_pairs(const Symmetry& symm);
-
-        void output_full_map_to_irreducible_sector(const int nat);
-        void output_sector_star();
-
+        std::vector<TC> get_Rs_from_BvK(const K_Vectors& kv)const;
+        std::vector<TC> get_Rs_from_adjacent_list(const UnitCell& ucell, Grid_Driver& gd, const Parallel_Orbitals& pv)const;
         //--------------------------------------------------------------------------------
 
         /// The sub functions to rotate matrices
         /// mode='H': H_12(R)=T^\dagger(V)H_1'2'(VR+O_1-O_2)T(V)
         /// mode='D': D_12(R)=T^T(V)D_1'2'(VR+O_1-O_2)T^*(V)
         template<typename Tdata>    // RI::Tensor type, blas
-        RI::Tensor<Tdata> rotate_atompair_serial(const RI::Tensor<Tdata>& t, const int isym,
+        RI::Tensor<Tdata> rotate_atompair_serial(const RI::Tensor<Tdata>& A, const int isym,
             const Atom& a1, const Atom& a2, const char mode, bool output = false)const;
+        template<typename Tdata>    // pointer type, blas
+        void rotate_atompair_serial(Tdata* TAT, const Tdata* A, const int& nw1, const int& nw2, const int isym,
+            const Atom& a1, const Atom& a2, const char mode)const;
         template<typename TR>    // HContainer type, pblas
         void rotate_atompair_parallel(const TR* Alocal_in, const int isym, const Atom* atoms, const Statistics& st,
             const Tap& ap_in, const Tap& ap_out, const char mode, const Parallel_Orbitals& pv, TR* Alocal_out, const bool output = false)const;
-        //--------------------------------------------------------------------------------
 
-        /// The sub functions judge special symmetry
-        void gen_symmetry_BvK(const Symmetry& symm, const Atom* atoms, const Lattice& lat, const Statistics& st, const TC bvk_period);
-        /// whether in 2D plain or not for each symmetry operation
-        std::vector<bool> in_plain(const ModuleSymmetry::Symmetry& symm, const ModuleBase::Matrix3& latvec)const;
+        template<typename Tdata>
+        RI::Tensor<Tdata> set_rotation_matrix(const Atom& a, const int& isym)const;
         //--------------------------------------------------------------------------------
 
         int nsym_ = 1;
@@ -210,30 +144,8 @@ namespace ModuleSymmetry
         std::vector<std::map<int, std::vector<std::complex<double>>>> Ms_;
 
         /// irreducible sector
-        /// irreducible atom pairs: [n_iap][(isym, ap=(iat1, iat2))]
-        // std::vector<std::map<int, Tap>> atompair_stars_;
+        Irreducible_Sector irs_;
 
-        ///The index range of the orbital matrix to be calculated: irreducible R in irreducible atom pairs
-        // (including R in other atom pairs that cannot rotate into R_stars_[irreducebule_ap])
-        std::map<Tap, std::set<TC>> irreducible_sector_;
-
-        // //[natoms*natoms](R, (isym, irreducible_R))
-        // std::vector<std::map<TC, std::pair<int, TC>>> full_map_to_irreducible_sector_;
-        // (abR) -> (isym, abR)
-        std::map<TapR, std::pair<int, TapR>, apR_less_func> full_map_to_irreducible_sector_;
-
-        // all the {abR}s , where the isym=0 one in each star forms the irreducible sector.
-        // [irreducible sector size][isym, ((ab),R)]
-        std::vector<std::map<int, TapR>> sector_stars_;
-
-        /// the direct lattice vector of {R|t}\tau-\tau' for each atoms and each symmetry operation. [natom][nsym]
-        std::vector<std::vector<TCdouble>> return_lattice_;
-
-        std::vector<int> invmap_;
-        std::vector<int> isymbvk_to_isym_;
-        std::vector<ModuleBase::Matrix3> bvk_gmatrix_;
-        std::vector<ModuleBase::Vector3<double>> bvk_gtrans_;
-        int bvk_nsym_;
     };
 }
