+++ source/module_ri/exx_symmetry/test/symmetry_rotation_test.cpp
