+++ source/module_ri/test/CMakeLists.txt
@@ -6,9 +6,3 @@ AddTest(
   LIBS base ${math_libs} device
   SOURCES dm_mixing_test.cpp ../Mix_DMk_2D.cpp ../Mix_Matrix.cpp
 )
-AddTest(
-  TARGET symmetry_rotation
-  LIBS base ${math_libs} device symmetry 
-  SOURCES symmetry_rotation_test.cpp ../symmetry_rotation.cpp ../../module_basis/module_ao/parallel_2d.cpp
-  ../../module_io/output.cpp
-)
